+++ src/ts/classes/StringCollection.ts
