+++ src/ts/classes/StringInfo.ts
