+++ src/ts/classes/StringState.ts
