+++ src/ts/static/Note.ts
