+++ src/ts/static/Strings.ts
@@ -1,6 +1,8 @@
-import { StringCollection } from "./StringCollection.js";
-import { StringInfo } from "./StringInfo.js";
-export { Strings };
+import { StringCollection } from "../classes/StringCollection.js"
+import { StringInfo } from "../classes/StringInfo.js"
+
+export { Strings }
+
 /**
  * Represents guitar string brands. Each brand of guitar string is a collection, represented as a StringCollection.
  */
@@ -10,8 +12,9 @@ class Strings {
      *
      * @returns {StringCollection} D'Addario Plain Steel strings.
      */
-    static dAddarioPlainSteel() {
-        let dAddarioPlainSteel = new StringCollection("D'Addario", "Plain Steel");
+    public static dAddarioPlainSteel(): StringCollection {
+        let dAddarioPlainSteel = new StringCollection("D'Addario", "Plain Steel")
+
         dAddarioPlainSteel.strings = ([
             new StringInfo(0.007, 0.00001085),
             new StringInfo(0.008, 0.00001418),
@@ -35,19 +38,23 @@ class Strings {
             new StringInfo(0.022, 0.00010722),
             new StringInfo(0.024, 0.0001276),
             new StringInfo(0.026, 0.00014975),
-        ]);
+        ])
+
         // Pass the brand and type down from the StringCollection to each StringInfo instance
-        dAddarioPlainSteel.setBrandForStrings();
-        dAddarioPlainSteel.setTypeForStrings();
-        return dAddarioPlainSteel;
+        dAddarioPlainSteel.setBrandForStrings()
+        dAddarioPlainSteel.setTypeForStrings()
+
+        return dAddarioPlainSteel
     }
+
     /**
      * Get the D'Addario XL Nickel Wound collection.
      *
      * @returns {StringCollection} D'Addario XL Nickel Wound strings.
      */
-    static dAddarioXLNickelWound() {
-        let dAddarioXLNickelWound = new StringCollection("D'Addario", "XL Nickel Wound");
+    public static dAddarioXLNickelWound(): StringCollection {
+        let dAddarioXLNickelWound = new StringCollection("D'Addario", "XL Nickel Wound")
+
         dAddarioXLNickelWound.strings = ([
             new StringInfo(0.017, 0.00005524),
             new StringInfo(0.018, 0.00006215),
@@ -82,10 +89,12 @@ class Strings {
             new StringInfo(0.072, 0.00094124),
             new StringInfo(0.074, 0.00098869),
             new StringInfo(0.08, 0.00115011),
-        ]);
+        ])
+
         // Pass the brand and type down from the StringCollection to each StringInfo instance
-        dAddarioXLNickelWound.setBrandForStrings();
-        dAddarioXLNickelWound.setTypeForStrings();
-        return dAddarioXLNickelWound;
+        dAddarioXLNickelWound.setBrandForStrings()
+        dAddarioXLNickelWound.setTypeForStrings()
+
+        return dAddarioXLNickelWound
     }
 }
