+++ src/ts/static/Utilities.ts
